feat(events): added event aggregate module
